splunk-custom-search-command-python/requirements.txt
splunk-sdk==1.6.14
splunklib==1.0.0
